Create README.md
Hello to all,
My name is Noah Brannon. I am a recent graduate from the University of Michigan - Dearborn with a degree in Economics.  
Recently I finished the Google Data Analytics Professional Certificate and I am continuing to learn how to use SQL and Python along with Tableau/PowerBI and uploading my projects to this repository. This will be an ongoing process. I am also looking for my first post-graduate job anything to do with data analysis preferably dealing with 
economics and/or finance.

The first project I uploaded was a simple Data Exploration project in SQL with Global COVID-19 Data from early 2020 to early March 2023. The code/syntax for that project 
is on this repository under the title "CovidProject.sql". For that project, I used Microsoft SQL Server Management Studio 19 on Windows 11.